Refactor send.email action
