+++ actions/src/send_email.py
@@ -1,155 +1,27 @@
-import os
-from smtplib import SMTP_SSL
-from email.header import Header
-from email.mime.application import MIMEApplication
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-from email.utils import formatdate
+from typing import Callable, Dict
+from email_api.email_api import EmailApi
 
 from st2common.runners.base_action import Action
 
 
 class SendEmail(Action):
-    def run(self, **kwargs):
+    def __init__(self, *args, config: Dict = None, **kwargs):
+        super().__init__(*args, config=config, **kwargs)
+        self._api: EmailApi = config.get("email_api", EmailApi())
+
+    def run(self, submodule: str, **kwargs):
         """
-        function that is run by stackstorm when an action is invoked
+        Dynamically dispatches to the method wanted
+        """
+        func: Callable = getattr(self, submodule)
+        return func(**kwargs)
+
+    def send_email(self, **kwargs):
+        """
+        Sends an email
         :param kwargs: arguments for openstack actions
         :return: (Status (Bool), Output <*>): tuple of action status (succeeded(T)/failed(F)) and the output
         """
         if "body" in kwargs and kwargs["body"]:
-            return self.send_email(**kwargs)
+            return self._api.send_email(self.config, **kwargs)
         return None
-
-    def load_template(self, prop, **kwargs):
-        """
-        Loads and returns a html template from its path
-        :param prop: Name of the property in kwargs containing the path
-                     to the template
-        :param kwargs: Arguments for the action
-        :return: (String) The contents of the file in utf-8 format
-        """
-        if prop in kwargs:
-            path = kwargs[prop]
-            if os.path.exists(path):
-                with open(path, "r", encoding="utf-8") as header_file:
-                    return header_file.read()
-            else:
-                raise FileNotFoundError(f"Template file located {path} not found")
-        else:
-            raise KeyError(f"The property {prop} does not appear in the configuration")
-
-    def load_smtp_account(self, **kwargs):
-        """
-        Loads and returns the smtp account data from the pack config
-        :param kwargs: Arguments for the action
-        :return: (Dictionary) SMTP account properties
-        """
-        accounts = self.config.get("smtp_accounts", None)
-        if not accounts:
-            raise ValueError(
-                f"'smtp_account' config value is required to send email, config={self.config}"
-            )
-        try:
-            key_value = {a["name"]: a for a in accounts}
-            account_data = key_value[kwargs["smtp_account"]]
-        except KeyError as exc:
-            raise KeyError(
-                f"The account {kwargs['smtp_account']} does not appear in the configuration"
-            ) from exc
-
-        return account_data
-
-    def attach_files(self, msg: MIMEMultipart, files):
-        """
-        Loads and adds attachments to an email message
-        :param msg: The message object for the email
-        :param files: List/Tuple of file paths of files to attach
-        :return:
-        """
-        for filepath in files:
-            filename = os.path.basename(filepath)
-            with open(filepath, "rb") as file:
-                part = MIMEApplication(file.read(), Name=filename)
-            part["Content-Disposition"] = f"attachment; filename={filename}"
-            msg.attach(part)
-
-    def send_email(self, **kwargs):
-        """
-        Send email
-        :param kwargs: email_to (String): Recipient Email,
-        email_from (String): Sender Email, subject (String): Email Subject,
-        header (String): filepath to header file,
-        footer (String): filepath to footer file,
-        body (String): body of email,
-        attachment (List): list of attachment filepaths,
-        smtp_account (String): email config to use,
-        admin_override (Boolean): send all emails to admin email - testing purposes,
-        admin_override_email (String): send to this email if admin_override enabled
-        :return:
-        """
-
-        account_data = self.load_smtp_account(**kwargs)
-
-        msg = MIMEMultipart()
-        msg["Subject"] = Header(kwargs["subject"], "utf-8")
-        msg["From"] = kwargs["email_from"]
-        msg["To"] = ", ".join(kwargs["email_to"])
-
-        if kwargs["admin_override"]:
-            msg["To"] = ", ".join(kwargs["admin_override_email"])
-
-        msg["Date"] = formatdate(localtime=True)
-
-        header = self.load_template("header", **kwargs)
-        footer = self.load_template("footer", **kwargs)
-
-        if kwargs["send_as_html"]:
-            msg.attach(
-                MIMEText(
-                    f"""\
-            <html>
-                <head>
-                <style>
-                    table, th, td {{ border: 1px solid black; border-collapse: collapse; }}
-                    th, td {{ padding: 8px; }}
-                </style>
-                </head>
-                <body>
-                    {header}
-                    {kwargs["body"]}
-                    {footer}
-                </body>
-            </html>
-            """,
-                    "html",
-                )
-            )
-        else:
-            msg.attach(
-                MIMEText(f"{header}\n{kwargs['body']}\n{footer}", "plain", "utf-8")
-            )
-
-        if kwargs["email_cc"]:
-            msg["Cc"] = ", ".join(kwargs["email_cc"])
-        self.attach_files(msg, kwargs["attachment_filepaths"] or tuple())
-
-        smtp = SMTP_SSL(account_data["server"], str(account_data["port"]), timeout=60)
-        smtp.ehlo()
-
-        if account_data.get("secure", True):
-            smtp.starttls()
-        if account_data.get("smtp_auth", True):
-            smtp.login(account_data["username"], account_data["password"])
-
-        if kwargs["admin_override"]:
-            email_to = kwargs["admin_override_email"]
-        else:
-            email_to = (
-                (kwargs["email_to"] + kwargs["email_cc"])
-                if kwargs["email_cc"]
-                else kwargs["email_to"]
-            )
-
-        smtp.sendmail(kwargs["email_from"], email_to, msg.as_string())
-        smtp.quit()
-        return (True, "Email sent")
